// Lesson47/recipe-app/src/test-utils/mocked-recipe.ts
import { Area } from '../modules/recipes/models/Area';
import { Category } from '../modules/recipes/models/Category';
import { RecipeDetails } from '../modules/recipes/models/RecipeDetails';

export const mockedRecipe: RecipeDetails = {
  idMeal: '52893',
  strMeal: 'Apple & Blackberry Crumble',
  strCategory: Category.Dessert,
  strArea: Area.British,
  strDrinkAlternate: undefined,
  strInstructions:
    'Heat oven to 190C/170C fan/gas 5. Tip the flour and sugar into a large bowl...',
  strMealThumb:
    'https://www.themealdb.com/images/media/meals/xvsurr1511719182.jpg',
  strTags: 'Pudding',
  strYoutube: 'https://www.youtube.com/watch?v=4vhcOwVBDO4',
  strIngredient1: 'Plain Flour',
  strIngredient2: 'Caster Sugar',
  strIngredient3: 'Butter',
  strIngredient4: 'Braeburn Apples',
  strIngredient5: 'Butter',
  strMeasure1: '120g',
  strMeasure2: '60g',
  strMeasure3: '60g',
  strMeasure4: '300g',
  strMeasure5: '30g',
  strSource:
    'https://www.bbcgoodfood.com/recipes/778642/apple-and-blackberry-crumble',
};
